Change ids to longs

+++ src/main/kotlin/com/x7ff/rl/replay/api/model/replay/DemolitionPlayer.kt
@@ -3,6 +3,6 @@ package com.x7ff.rl.replay.api.model.replay
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class DemolitionPlayer(
-    @JsonProperty("id") val id: Int,
+    @JsonProperty("id") val id: Long,
     @JsonProperty("name") val name: String
 )
